autoscaling nodegroup, support for minimum size=zero
Use the stackpoint.io Node.Group = autoscaling field and the kubernetes
stackpointio/node_group=autoscaling label to indicate the autoscaling
group.  Don't manage other nodes.

Circumvent the various constructors and checks that rely on having
an initial node that can be used as a template. This fails when
the autoscaling group has a zero lower bound.  Instead, simply assume
that one node is needed and begin scaling from there.